Make pip do not hang while installing
Pip dependency resolver has had a bug fixed in the past. Apparently this
fix makes dependency resolution more correct but it requires downloading
of lots and lots of versions of the same package, causing pip install to
take a very long time. This uses the old dependency resolver to be able
to move forward. This is just a temporary measure until a better
solution can be found.