Update environment variables (#72)
* Revert "Fix testing requirements"

This reverts commit 83ec77368917330b9b7f403af093f0a3638baf05.

* Make pip do not hang while installing

Pip dependency resolver has had a bug fixed in the past. Apparently this
fix makes dependency resolution more correct but it requires downloading
of lots and lots of versions of the same package, causing pip install to
take a very long time. This uses the old dependency resolver to be able
to move forward. This is just a temporary measure until a better
solution can be found.

* Fixing supported version

One of our dependencies (environs) no longer supports 3.5. So, removing
support for 3.5. Also, ipython only supports 3.6 for 7.16.

* Use 3.6 environment

* Fix lint

* Remove remaining metric references

* Update environment variables

* Fix reqs

* Support legacy environment variables

* Fix lint